+++ packages/mongodb-memory-server-core/src/__tests__/MongoMemoryServer.test.ts
@@ -77,6 +77,10 @@ describe('MongoMemoryServer', () => {
         instance: {
           storageEngine: 'ephemeralForTest',
         },
+        binary: {
+          // 7.0 removed "ephemeralForTest", this test is explicitly for that engine
+          version: '6.0.14',
+        },
       });
 
       utils.assertion(!utils.isNullOrUndefined(mongoServer.instanceInfo));
@@ -259,6 +263,10 @@ describe('MongoMemoryServer', () => {
         instance: {
           storageEngine: 'ephemeralForTest',
         },
+        binary: {
+          // 7.0 removed "ephemeralForTest", this test is explicitly for that engine
+          version: '6.0.14',
+        },
       });
 
       utils.assertion(!utils.isNullOrUndefined(mongoServer.instanceInfo));
@@ -368,9 +376,6 @@ describe('MongoMemoryServer', () => {
         auth: {
           enable: false,
         },
-        instance: {
-          storageEngine: 'ephemeralForTest',
-        },
       });
 
       utils.assertion(!utils.isNullOrUndefined(mongoServer.instanceInfo));
@@ -452,6 +457,142 @@ describe('MongoMemoryServer', () => {
       await expect(mongoServer2.start()).rejects.toThrow(
         'Cannot start because "instance.mongodProcess" is already defined!'
       );
+    });
+
+    describe('instance.portGeneration', () => {
+      it('should use a predefined port if "opts.instance.portGeneration" is "false"', async () => {
+        const predefinedPort = 30001;
+
+        const mongoServer = new MongoMemoryServer({
+          instance: { port: predefinedPort, portGeneration: false },
+        });
+        const newPortSpy = jest
+          // @ts-expect-error "getNewPort" is protected
+          .spyOn(mongoServer, 'getNewPort')
+          .mockImplementation(() => fail('Expected this function to not be called'));
+
+        await mongoServer.start();
+
+        expect(newPortSpy).not.toHaveBeenCalled();
+        // @ts-expect-error "_instanceInfo" is protected
+        expect(mongoServer._instanceInfo!.port).toStrictEqual(predefinedPort);
+
+        await mongoServer.stop();
+      });
+
+      it('should Error if a predefined port is already in use if "opts.instance.portGeneration" is "false"', async () => {
+        const predefinedPort = 30002;
+
+        const newPortSpy = jest
+          // @ts-expect-error "getNewPort" is protected
+          .spyOn(MongoMemoryServer.prototype, 'getNewPort')
+          .mockImplementation(() => fail('Expected this function to not be called'));
+
+        jest.spyOn(console, 'warn').mockImplementationOnce(() => void 0);
+
+        const mongoServer1 = new MongoMemoryServer({
+          instance: { port: predefinedPort, portGeneration: false },
+        });
+
+        const mongoServer2 = new MongoMemoryServer({
+          instance: { port: predefinedPort, portGeneration: false },
+        });
+
+        await mongoServer1.start();
+
+        await expect(() => mongoServer2.start()).rejects.toMatchSnapshot();
+
+        expect(newPortSpy).not.toHaveBeenCalled();
+        // @ts-expect-error "_instanceInfo" is protected
+        expect(mongoServer1._instanceInfo!.port).toStrictEqual(predefinedPort);
+        expect(console.warn).toHaveBeenCalledTimes(1);
+
+        await mongoServer1.stop();
+      });
+
+      it('should generate a new port if the predefined port is already in use and "opts.instance.portGeneration" is "true"', async () => {
+        const predefinedPort = 30003;
+
+        const newPortSpy = jest
+          // @ts-expect-error "getNewPort" is protected
+          .spyOn(MongoMemoryServer.prototype, 'getNewPort');
+
+        const mongoServer1 = new MongoMemoryServer({
+          instance: { port: predefinedPort, portGeneration: false },
+        });
+
+        const mongoServer2 = new MongoMemoryServer({
+          instance: { port: predefinedPort, portGeneration: true },
+        });
+
+        await mongoServer1.start();
+
+        expect(newPortSpy).not.toHaveBeenCalled();
+
+        await mongoServer2.start();
+
+        expect(newPortSpy).toHaveBeenCalledTimes(1);
+
+        await mongoServer1.stop();
+        await mongoServer2.stop();
+      });
+
+      it('should overwrite "opts.instance.portGeneration" if "forceSamePort" is set ("forceSamePort true" case)', async () => {
+        const predefinedPort = 30004;
+
+        const newPortSpy = jest
+          // @ts-expect-error "getNewPort" is protected
+          .spyOn(MongoMemoryServer.prototype, 'getNewPort')
+          .mockImplementation(() => fail('Expected this function to not be called'));
+
+        jest.spyOn(console, 'warn').mockImplementationOnce(() => void 0);
+
+        const mongoServer1 = new MongoMemoryServer({
+          instance: { port: predefinedPort, portGeneration: false },
+        });
+
+        const mongoServer2 = new MongoMemoryServer({
+          instance: { port: predefinedPort, portGeneration: true },
+        });
+
+        await mongoServer1.start();
+
+        await expect(() => mongoServer2.start(true)).rejects.toMatchSnapshot();
+
+        expect(newPortSpy).not.toHaveBeenCalled();
+        // @ts-expect-error "_instanceInfo" is protected
+        expect(mongoServer1._instanceInfo!.port).toStrictEqual(predefinedPort);
+        expect(console.warn).toHaveBeenCalledTimes(1);
+
+        await mongoServer1.stop();
+      });
+
+      it('should overwrite "opts.instance.portGeneration" if "forceSamePort" is set ("forceSamePort false" case)', async () => {
+        const predefinedPort = 30005;
+
+        const newPortSpy = jest
+          // @ts-expect-error "getNewPort" is protected
+          .spyOn(MongoMemoryServer.prototype, 'getNewPort');
+
+        const mongoServer1 = new MongoMemoryServer({
+          instance: { port: predefinedPort, portGeneration: false },
+        });
+
+        const mongoServer2 = new MongoMemoryServer({
+          instance: { port: predefinedPort, portGeneration: false },
+        });
+
+        await mongoServer1.start();
+
+        expect(newPortSpy).not.toHaveBeenCalled();
+
+        await mongoServer2.start(false);
+
+        expect(newPortSpy).toHaveBeenCalledTimes(1);
+
+        await mongoServer1.stop();
+        await mongoServer2.stop();
+      });
     });
   });
 
@@ -623,22 +764,6 @@ describe('MongoMemoryServer', () => {
       expect(cleanupSpy).toHaveBeenCalledWith({ doCleanup: true } as utils.Cleanup);
     });
 
-    it('should not support boolean arguments', async () => {
-      const mongoServer = new MongoMemoryServer();
-
-      try {
-        await mongoServer.stop(
-          // @ts-expect-error Testing a non-existing overload
-          true
-        );
-        fail('Expected to fail');
-      } catch (err) {
-        expect(err).toBeInstanceOf(Error);
-        assertIsError(err);
-        expect(err.message).toMatchSnapshot();
-      }
-    });
-
     it('should call cleanup and pass-through cleanup options', async () => {
       const mongoServer = new MongoMemoryServer();
 
@@ -789,22 +914,6 @@ describe('MongoMemoryServer', () => {
       expect(await utils.statPath(dbPath)).toBeUndefined();
       expect(mongoServer.state).toEqual(MongoMemoryServerStates.new);
       expect(mongoServer.instanceInfo).toBeUndefined();
-    });
-
-    it('should not support boolean arguments', async () => {
-      const mongoServer = new MongoMemoryServer();
-
-      try {
-        await mongoServer.cleanup(
-          // @ts-expect-error Testing a non-existing overload
-          true
-        );
-        fail('Expected to fail');
-      } catch (err) {
-        expect(err).toBeInstanceOf(Error);
-        assertIsError(err);
-        expect(err.message).toMatchSnapshot();
-      }
     });
 
     it('should properly cleanup with tmpDir with default no force (new)', async () => {
@@ -1181,6 +1290,75 @@ describe('MongoMemoryServer', () => {
       expect(server.instanceInfo?.storageEngine).toStrictEqual('wiredTiger');
 
       await server.stop();
+    });
+  });
+
+  describe('asyncDispose', () => {
+    it('should work by default', async () => {
+      jest.spyOn(MongoMemoryServer.prototype, 'start');
+      jest.spyOn(MongoMemoryServer.prototype, 'stop');
+      let outer;
+      // would like to test this, but jest seemingly does not support spying on symbols
+      // jest.spyOn(MongoMemoryServer.prototype, Symbol.asyncDispose);
+      {
+        await using server = await MongoMemoryServer.create();
+        // use the value and test that it actually runs, as "getUri" will throw is not in "running" state
+        server.getUri();
+        // reassignment still calls dispose at the *current* scope
+        outer = server;
+      }
+      // not "stopped" because of cleanup
+      expect(outer.state).toStrictEqual(MongoMemoryServerStates.new);
+      expect(MongoMemoryServer.prototype.start).toHaveBeenCalledTimes(1);
+      expect(MongoMemoryServer.prototype.stop).toHaveBeenCalledTimes(1);
+      // expect(MongoMemoryServer.prototype[Symbol.asyncDispose]).toHaveBeenCalledTimes(1);
+    });
+
+    it('should be able to be disabled', async () => {
+      jest.spyOn(MongoMemoryServer.prototype, 'start');
+      jest.spyOn(MongoMemoryServer.prototype, 'stop');
+      let outer;
+      // would like to test this, but jest seemingly does not support spying on symbols
+      // jest.spyOn(MongoMemoryServer.prototype, Symbol.asyncDispose);
+      {
+        await using server = await MongoMemoryServer.create({ dispose: { enabled: false } });
+        // use the value and test that it actually runs, as "getUri" will throw is not in "running" state
+        server.getUri();
+        // reassignment still calls dispose at the *current* scope
+        outer = server;
+      }
+      expect(outer.state).toStrictEqual(MongoMemoryServerStates.running);
+      expect(MongoMemoryServer.prototype.start).toHaveBeenCalledTimes(1);
+      expect(MongoMemoryServer.prototype.stop).toHaveBeenCalledTimes(0);
+      // expect(MongoMemoryServer.prototype[Symbol.asyncDispose]).toHaveBeenCalledTimes(1);
+      await outer.stop();
+      // not "stopped" because of cleanup
+      expect(outer.state).toStrictEqual(MongoMemoryServerStates.new);
+    });
+
+    it('should be able to set custom cleanup', async () => {
+      jest.spyOn(MongoMemoryServer.prototype, 'start');
+      jest.spyOn(MongoMemoryServer.prototype, 'stop');
+      let outer;
+      // would like to test this, but jest seemingly does not support spying on symbols
+      // jest.spyOn(MongoMemoryServer.prototype, Symbol.asyncDispose);
+      {
+        await using server = await MongoMemoryServer.create({
+          dispose: { cleanup: { doCleanup: false } },
+        });
+        // use the value and test that it actually runs, as "getUri" will throw is not in "running" state
+        server.getUri();
+        // reassignment still calls dispose at the *current* scope
+        outer = server;
+      }
+      // not "stopped" because of cleanup
+      expect(outer.state).toStrictEqual(MongoMemoryServerStates.stopped);
+      expect(MongoMemoryServer.prototype.start).toHaveBeenCalledTimes(1);
+      expect(MongoMemoryServer.prototype.stop).toHaveBeenCalledTimes(1);
+      // expect(MongoMemoryServer.prototype[Symbol.asyncDispose]).toHaveBeenCalledTimes(1);
+      await outer.cleanup({ doCleanup: true });
+      // not "stopped" because of cleanup
+      expect(outer.state).toStrictEqual(MongoMemoryServerStates.new);
     });
   });
 });