+++ packages/mongodb-memory-server-core/src/util/MongoBinaryDownloadUrl.ts
@@ -481,14 +481,14 @@ export class MongoBinaryDownloadUrl implements MongoBinaryDownloadUrlOpts {
           19: '18.04',
           20: '20.04',
           21: '22.04',
+          22: '24.04',
         };
 
         ubuntuOS = {
           os: 'linux',
           dist: 'ubuntu',
           release:
-            // TODO: next-minor: change default mint release to 21
-            mintToUbuntuRelease[parseInt(os.release.split('.')[0])] || mintToUbuntuRelease[20],
+            mintToUbuntuRelease[parseInt(os.release.split('.')[0])] || mintToUbuntuRelease[21],
         };
       }
 
@@ -503,13 +503,13 @@ export class MongoBinaryDownloadUrl implements MongoBinaryDownloadUrlOpts {
 
         // untangle elemenatary versioning from hell https://en.wikipedia.org/wiki/Elementary_OS#Development
         const [elementaryMajor, elementaryMinor] = os.release.split('.').map((el) => parseInt(el));
+        // versions below 5.0 were named 0.X, and so use the minor version if major is 0
         const realMajor = elementaryMajor || elementaryMinor;
 
         ubuntuOS = {
           os: 'linux',
           dist: 'ubuntu',
-          // TODO: next-minor: change default elemenatry release to 7
-          release: elementaryToUbuntuRelease[realMajor] || elementaryToUbuntuRelease[6],
+          release: elementaryToUbuntuRelease[realMajor] || elementaryToUbuntuRelease[7],
         };
       }
     }
@@ -588,7 +588,7 @@ export class MongoBinaryDownloadUrl implements MongoBinaryDownloadUrlOpts {
        * Highest ubuntu year supported by mongodb binaries
        * @see https://www.mongodb.com/download-center/community/releases/archive
        */
-      const highestUbuntuYear = 22; // 22 is the highest supported as of mongodb 7.0.4
+      const highestUbuntuYear = 22; // 22 is the highest supported as of mongodb 7.0.11
 
       if (ubuntuYear > highestUbuntuYear) {
         log(