+++ packages/mongodb-memory-server-core/src/util/getport/index.ts
@@ -1,5 +1,3 @@
-import resolveConfig, { ResolveConfigVariables, envToBool } from '../resolveConfig';
-import * as crypto from 'node:crypto';
 import * as net from 'node:net';
 import debug from 'debug';
 
@@ -16,8 +14,6 @@ interface IPortsCache {
   timeSet: undefined | number;
   /** The ports that were tried */
   ports: Set<number>;
-  /** Store last used number, reduces amount of tries needed */
-  lastNumber: number;
 }
 
 /**
@@ -32,15 +28,14 @@ const PORTS_CACHE_CLEAN_TIME = 1000 * 10;
 const PORTS_CACHE: IPortsCache = {
   timeSet: undefined,
   ports: new Set(),
-  lastNumber: MIN_PORT,
 };
 
 /** Max default tries before giving up */
 const MAX_DEFAULT_TRIES = 10;
 
 /**
- * Try to get a free port
- * @param firstPort The first port to try or empty for semi-random port
+ * Try to get a free port.
+ * @param firstPort The first port to try or empty for a random port
  * @param max_tries maximum amount of tries to get a port, default to {@link MAX_DEFAULT_TRIES}
  * @returns A valid free port
  * @throws if "max_tries" is exceeded
@@ -49,8 +44,8 @@ export async function getFreePort(
   firstPort?: number,
   max_tries: number = MAX_DEFAULT_TRIES
 ): Promise<number> {
-  // Get a random value from crypto to use as first port if none is given
-  firstPort = firstPort || validPort(crypto.randomInt(MIN_PORT, MAX_PORT + 1));
+  // use "0" as a fallback to use net0listen, which generates a random free port
+  firstPort = firstPort || 0;
 
   // clear ports cache after some time, but not on an interval
   if (PORTS_CACHE.timeSet && Date.now() - PORTS_CACHE.timeSet > PORTS_CACHE_CLEAN_TIME) {
@@ -60,22 +55,12 @@ export async function getFreePort(
     PORTS_CACHE.timeSet = Date.now();
   }
 
-  const exp_net0listen = envToBool(resolveConfig(ResolveConfigVariables.EXP_NET0LISTEN));
-  log('EXP_NET0LISTEN', exp_net0listen);
-
   let tries = 0;
   while (tries <= max_tries) {
     tries += 1;
 
-    let nextPort: number;
-
-    if (exp_net0listen) {
-      // "0" means to use ".listen" random port
-      nextPort = tries === 1 ? firstPort : 0;
-    } else {
-      // use "startPort" at first try, otherwise increase from last number
-      nextPort = tries === 1 ? firstPort : validPort(PORTS_CACHE.lastNumber + tries);
-    }
+    // "0" means to use have ".listen" use a random port
+    const nextPort = tries === 1 ? firstPort : 0;
 
     // try next port, because it is already in the cache
     // unless port is "0" which will use "net.listen(0)"
@@ -84,15 +69,15 @@ export async function getFreePort(
     }
 
     PORTS_CACHE.ports.add(nextPort);
-    // only set "lastNumber" if the "nextPort" was not in the cache
-    PORTS_CACHE.lastNumber = nextPort;
 
     const triedPort = await tryPort(nextPort);
 
-    // returned port can be different than the "nextPort" (if EXP_NET0LISTEN)
-    PORTS_CACHE.ports.add(nextPort);
-
     if (triedPort > 0) {
+      log('getFreePort: found free port', triedPort);
+
+      // returned port can be different than the "nextPort" (if net0listen)
+      PORTS_CACHE.ports.add(nextPort);
+
       return triedPort;
     }
   }
@@ -103,8 +88,8 @@ export async function getFreePort(
 export default getFreePort;
 
 /**
- * Check that input number is within range of {@link MIN_PORT} and {@link MAX_PORT}
- * If more than {@link MAX_PORT}, wrap around, if less than {@link MIN_PORT} use {@link MIN_PORT}
+ * Ensure that input number is within range of {@link MIN_PORT} and {@link MAX_PORT}.
+ * If more than {@link MAX_PORT}, wrap around, if less than {@link MIN_PORT} use {@link MIN_PORT}.
  * @param port The Number to check
  * @returns A Valid number in port range
  */
@@ -115,9 +100,9 @@ export function validPort(port: number): number {
 }
 
 /**
- * Try a given port
+ * Try a given port.
  * @param port The port to try
- * @returns "true" if the port is not in use, "false" if in use
+ * @returns the port if successful, "-1" in case of `EADDRINUSE`, all other errors reject
  * @throws The error given if the code is not "EADDRINUSE"
  */
 export function tryPort(port: number): Promise<number> {
@@ -147,12 +132,11 @@ export function tryPort(port: number): Promise<number> {
 }
 
 /**
- * Reset the {@link PORTS_CACHE} to its initial state
+ * Reset the {@link PORTS_CACHE} to its initial state.
  *
- * This function is meant for debugging and testing purposes only
+ * This function is meant for debugging and testing purposes only.
  */
 export function resetPortsCache() {
-  PORTS_CACHE.lastNumber = MIN_PORT;
   PORTS_CACHE.timeSet = undefined;
   PORTS_CACHE.ports.clear();
 }