+++ website/src/pages/index.js
@@ -10,18 +10,28 @@ import styles from './styles.module.css';
 const features = [
   {
     title: 'Easy to Use',
-    description: <>mongodb-memory-server was build with "easy-to-use" in mind</>,
-  },
-  {
-    title: 'Focus on What Matters',
-    description: <>mongodb-memory-server can be quickly added to projects on many systems</>,
-  },
-  {
-    title: 'Supports Many Systems',
     description: (
       <>
-        mongodb-memory-server supports many different systems without problems (exceptions are some
-        linux distributions)
+        mongodb-memory-server was build with "easy-to-use" in mind and is in most common systems
+        zero-configuration. (except setting a custom mongodb version)
+      </>
+    ),
+  },
+  {
+    title: 'Supports all MongoDB supports systems',
+    description: (
+      <>
+        mongodb-memory-server supporst all systems there are mongodb binaries are for, including
+        some unofficial environments where some other binary works.
+      </>
+    ),
+  },
+  {
+    title: 'Supports both Single-Instance and Replica Sets',
+    description: (
+      <>
+        mongodb-memory-server can be quickly used with single mongodb instances or spin up a entire
+        replicaset.
       </>
     ),
   },
@@ -59,7 +69,7 @@ function Home() {
                 'button button--outline button--secondary button--lg',
                 styles.getStarted
               )}
-              to={useBaseUrl('docs/')}
+              to={useBaseUrl('docs/guides/quick-start-guide')}
             >
               Get Started
             </Link>